requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
defusedxml==0.8.0rc2
Django==5.0.2
django-channels==0.7.0
django-restframework==0.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
kombu==5.3.5
oauthlib==3.2.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
python3-openid==3.2.0
pytz==2024.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.2
